api/views.py
from api.serializers import MessageSerializer
from api.models import Message
from django.shortcuts import render
from rest_framework import generics

class MessageListCreateAPIView(generics.ListCreateAPIView):
    queryset = Message.objects.all()
    serializer_class = MessageSerializer

class MessageDestroyAPIView(generics.DestroyAPIView):
    queryset = Message.objects.all()
    serializer_class = MessageSerializer
