requirements.txt
asgiref==3.3.1
dj-static==0.0.6
Django==3.1.5
djangorestframework==3.12.2
gunicorn==20.0.4
pytz==2020.5
sqlparse==0.4.1
static3==0.7.0
